feat: modification des uses case agriculteur (MettreAJourStockUseCase.php,ConsulterCommandesRecuesUseCaseTest.php,ValiderCommandeUseCase.php,ConsulterCommandesRecuesUseCase.php  )et leurs differents tests

+++ app/Application/Agriculteur/UseCase/SupprimerProduitUseCase.php
@@ -13,14 +13,20 @@ class SupprimerProduitUseCase
         $this->produitRepository = $produitRepository;
     }
 
-    public function execute(string $produitId): void
-    {
-        $produit = $this->produitRepository->findById($produitId);
+   public function execute(string $produitId): void
+{
+    $produit = $this->produitRepository->findById($produitId);
 
-        if (!$produit) {
-            throw new \Exception("Produit introuvable.");
-        }
-
-        $this->produitRepository->delete($produitId);
+    if (!$produit) {
+        throw new \Exception('Produit introuvable.');
     }
+
+    if ($produit->aDesCommandesEnCours()) {
+        throw new \DomainException(
+            'Impossible de supprimer ce produit.'
+        );
+    }
+
+    $this->produitRepository->delete($produitId);
+}
 }
